+++ SignNow.Net.Test/TestData/FakeModels/SignNowDocumentFaker.cs
@@ -1,4 +1,3 @@
-using System.IO;
 using Bogus;
 using SignNow.Net.Model;